Moving plot backend modules to a separate backends package
