+++ silx/gui/plot/backends/BackendBase.py
