+++ silx/gui/plot/backends/_matplotlib.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 # /*##########################################################################
 #
-# Copyright (c) 2016 European Synchrotron Radiation Facility
+# Copyright (c) 2016-2017 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -46,7 +46,7 @@ if 'matplotlib' in sys.modules:
         'matplotlib already loaded, setting its backend may not work')
 
 
-from .. import qt
+from ... import qt
 
 import matplotlib
